Handle an email address in the Slack Topic
What is happening: Behind the scenes, transparently to us, Slack is
adding a 'mailto:' link on the email address. The approach to get around
that is to regex match the string:
'<mailto:foo@example.com|foo@example.com>' and then capture the email
address. Using regex substitution, replace the string with the email
address and continue like that.