lambda: set slack \@handle
It's nice to be able to set \@handle in Slack topic because it allows
person to be notified by Slack when they go oncall, and because other
people can click that handle to go to DM or get info about the oncall
person.

This commit makes this the new behavior in lambda/main.py. Because this
behavior requires getting Slack user from the Slack API, and because Bot
users don't have that permission, this commit also updates README.md
with instructions on creating an App user with sufficient permissions.